+++ packages/solid-query-persist-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tanstack/solid-query-persist-client
 
+## 5.90.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.11
+  - @tanstack/solid-query@5.90.15
+
 ## 5.90.14
 
 ### Patch Changes